Update README.md
This Project outlines the development of a PageRank algorithm for link analysis using Apache Spark. The algorithm is applied to a dataset of customer reviews and product information from an e-commerce platform (Amazon products). The goal is to calculate product rankings based on weighted links between products, facilitating personalized product recommendations.